+++ AppIntents/Intents/SetPlayerValueIntent.swift
@@ -28,9 +28,13 @@ enum PlayerValueError: Error, CustomLocalizedStringResourceConvertible {
 
 @available(iOS 17.0, macOS 14.0, *)
 struct SetPlayerValueIntent: AppIntent {
-    static var openAppWhenRun: Bool { false }
+    static var openAppWhenRun: Bool {
+        false
+    }
 
-    static var allowedItemTypes: [OpenHABItem.ItemType] { [.player] }
+    static var allowedItemTypes: [OpenHABItem.ItemType] {
+        [.player]
+    }
 
     static var parameterSummary: some ParameterSummary {
         Summary("Send \(\.$action) to \(\.$itemEntity)") {