+++ OpenHABCore/Sources/OpenHABCore/Util/Collection+SafeAccess.swift
@@ -12,7 +12,7 @@
 import Foundation
 
 public extension Collection {
-    // Returns the element at the specified index if it is within bounds, otherwise nil.
+    /// Returns the element at the specified index if it is within bounds, otherwise nil.
     subscript(safe index: Index) -> Element? {
         indices.contains(index) ? self[index] : nil
     }